fix: Move and add observe functions in references.go; use observe for user references
